Anthony's spring cleaning

+++ src/main/java/org/usfirst/frc/team3042/robot/subsystems/DrivetrainEncoders.java
@@ -10,7 +10,7 @@ import com.ctre.phoenix.motorcontrol.can.TalonSRX;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import edu.wpi.first.wpilibj.smartdashboard.SendableRegistry;
 
-/** DrivetrainEncoders ***********************************************************
+/** DrivetrainEncoders ********************************************************
  * The encoders for the drivetrain.
  */
 public class DrivetrainEncoders extends Subsystem {
@@ -41,11 +41,9 @@ public class DrivetrainEncoders extends Subsystem {
 		reset();
 	}
 	private void initEncoder(TalonSRX encoder, boolean sensorPhase) {
-		encoder.configSelectedFeedbackSensor(FeedbackDevice.QuadEncoder, 
-				PIDIDX, TIMEOUT);
-		encoder.setStatusFramePeriod(StatusFrameEnhanced.Status_2_Feedback0, 
-				FRAME_RATE, TIMEOUT);
-		encoder.setSensorPhase(sensorPhase); 	// affects closed-loop mode
+		encoder.configSelectedFeedbackSensor(FeedbackDevice.QuadEncoder, PIDIDX, TIMEOUT);
+		encoder.setStatusFramePeriod(StatusFrameEnhanced.Status_2_Feedback0, FRAME_RATE, TIMEOUT);
+		encoder.setSensorPhase(sensorPhase); // affects closed-loop mode
 	}
 	
 	/** initDefaultCommand ****************************************************
@@ -57,10 +55,10 @@ public class DrivetrainEncoders extends Subsystem {
 
 	/** reset *****************************************************************/
 	public void reset() {
-		int leftCounts = leftEncoder.getSelectedSensorPosition(PIDIDX);
+		int leftCounts = (int)(leftEncoder.getSelectedSensorPosition(PIDIDX));
 		leftPositionZero = countsToRev(leftCounts);
 		
-		int rightCounts = rightEncoder.getSelectedSensorPosition(PIDIDX);
+		int rightCounts = (int)(rightEncoder.getSelectedSensorPosition(PIDIDX));
 		rightPositionZero = countsToRev(rightCounts);
 	}
 	public void setToZero() {
@@ -73,22 +71,22 @@ public class DrivetrainEncoders extends Subsystem {
 	 * Speed returns counts per 100ms and is converted to RPM
 	 */
 	public double getLeftPosition() {
-		int counts = leftEncoder.getSelectedSensorPosition(PIDIDX);
+		int counts = (int)(leftEncoder.getSelectedSensorPosition(PIDIDX));
 		return countsToRev(counts) - leftPositionZero;
 	}
 	public double getRightPosition() {
-		int counts = rightEncoder.getSelectedSensorPosition(PIDIDX);
+		int counts = (int)(rightEncoder.getSelectedSensorPosition(PIDIDX));
 		return countsToRev(counts) - rightPositionZero;
 	}
 	private double countsToRev(int counts) {
 		return (double)counts / COUNTS_PER_REVOLUTION;
 	}
 	public double getLeftSpeed() {
-		int cp100ms = leftEncoder.getSelectedSensorVelocity(PIDIDX);
+		int cp100ms = (int)(leftEncoder.getSelectedSensorVelocity(PIDIDX));
 		return cp100msToRPM(cp100ms);
 	}
 	public double getRightSpeed() {
-		int cp100ms = rightEncoder.getSelectedSensorVelocity(PIDIDX);
+		int cp100ms = (int)(rightEncoder.getSelectedSensorVelocity(PIDIDX));
 		return cp100msToRPM(cp100ms);
 	}
 	private double cp100msToRPM(int cp100ms) {
